+++ localfullnode2/src/main/java/one/inve/localfullnode2/store/EventStoreDependent.java
@@ -9,7 +9,7 @@ public interface EventStoreDependent {
 
 	int getShardCount();
 
-	int getCreatorId();
+	long getCreatorId();
 
 	BlockingQueue<EventBody> getEventSaveQueue();
 }
